Delete legacy json format
